[Feat]#24-화면전환,지역명함수생성
