+++ TodayWeather/TodayWeather/Presentation/CurrentWeather.swift
@@ -6,10 +6,12 @@
 //
 
 import UIKit
+import CoreLocation
 
 class CurrentWeather {
     static let shared = CurrentWeather()
     static var id = 0
+    private let CDM = CoreDataManager()
     
     func weatherImage(weather : Int) -> UIImage {
         let weatherState = WeatherModel(id: weather)
@@ -38,4 +40,26 @@ class CurrentWeather {
             return UIColor(named: "fewCloudyBackground") ?? UIColor()
         }
     }
+    
+    func reverseGeocode(latitude: Double, longitude: Double, save: Bool, completion : @escaping (Result<String, Error>) -> Void) {
+           let location = CLLocation(latitude: latitude, longitude: longitude)
+            let geocoder: CLGeocoder = CLGeocoder()
+            let local: Locale = Locale(identifier: "en-US")
+           geocoder.reverseGeocodeLocation(location, preferredLocale: local) { [weak self](placemarks, error) in
+               if let error = error {
+                   completion(.failure(error))
+                   return
+               }
+               
+               guard let placemark = placemarks?.first else {
+                   print("No placemark found")
+                   return
+               }
+               let name = (placemark.locality ?? "") + ", " + (placemark.country ?? "")
+               if save == true {
+                   self?.CDM.saveData(Data: locationData(latitude: latitude, longitude: longitude, locName: name))
+               }
+               completion(.success(name))
+           }
+       }
 }
